+++ src/features/show/actions.d.ts
@@ -1,3 +1,4 @@
 import type { AppThunk } from '~/store/reducers';
 
+export const authorizeIfAndOnlyIfHasStartTime: () => AppThunk;
 export const clearStartTime: () => AppThunk;